requirements.txt
flet
psycopg2-binary
bcrypt
reportlab